ConfigEntryNotReady in init
