+++ homeassistant/components/ipma/const.py
@@ -6,3 +6,6 @@ DOMAIN = "ipma"
 HOME_LOCATION_NAME = "Home"
 
 ENTITY_ID_SENSOR_FORMAT_HOME = f"{WEATHER_DOMAIN}.ipma_{HOME_LOCATION_NAME}"
+
+DATA_LOCATION = "location"
+DATA_API = "api"
